Add main HTML structure, CSS styles, and JS script
I added the main HTML structure of the Etch-a-Sketch project by creating
a header, a main container where the grid cells are going to be placed,
and a footer.

As for the CSS styles, I once again utilized Josh W. Comeau's CSS reset
and added basic design styles and flexbox layout to the content. This
will be polished further once the functionality of the script is
complete.

For the JS script, I initialized two function expressions: one for
calculating container width, and one for generating the grid cells. I
added parameters to these functions so I can flexibly utilize them
instead of manually changing the values.