Modify HTML structure, CSS styles, and JS script
I placed the reset and generate buttons inside a div to utilize flexbox.

I removed and added some styles in CSS. Instead of using gap, I Instead
used outline so that it is not hard on the eyes.

I removed the calculateContainerWidth function and instead calculate the
dimensions of the grid cell dynamically. I also followed Airbnb's
JavaScript style guide to clean up some of my code.