Implement modifications in JavaScript, HTML, CSS
I implemented the two optional challenge modifications for this project.
To generate the random cell colors, I utilized Math.random function and
multiplied it with the max RGB color of 255. To use this effect, the
user must hold the control key while hovering the cells.

To progressively darken the cells, I utilized the getComputedStyle
function to grab the background color of the cell, parse it through a
string manipulation function, and used the color values to either darken
a white cell into black or darken a colored cell into black. To use this
effect, the user must hold the shift key while hovering the cells.

I implemented an eraser hover function that removes the added
background color of the cell, essentially reverting it back to its
original color. I also renamed the reset button to clean button and its
function is similar to the eraser but it instead erases the whole grid
giving the user a clean slate to work with.