Add enable outline and finish HTML and CSS styles
I added a checkbox in the HTML to allow users to enable grid outlines of
the sketchpad. I utilized an event listener to dynamically enable or
disable the grid outline. Additionally, I passed the value of the
checkbox to the generateGrid function so that when a new grid is defined
it still has the grid outline enabled/disabled.

I designed the web project based on an Etch-a-Sketch by integrating its
colors and rounding the corners of the containers. I also chose a font
family that looks like a children's handwriting so that it feels like a
children's toy.